requirements.txt
python-dotenv
requests
bs4
markdownify
markdown
rich
pyyaml
art
prompt_toolkit
